clk: fix type cast warning
Reviewed-on: http://git.am.freescale.net:8181/37122
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>

+++ drivers/clk/clk-qoriq.c
@@ -127,7 +127,7 @@ static void __init core_mux_init(struct device_node *np)
 	init.name = clk_name;
 	init.ops = &cmux_ops;
 	init.parent_names = parent_names;
-	init.num_parents = count;
+	init.num_parents = (u8)count;
 	init.flags = 0;
 	cmux_clk->hw.init = &init;
